requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Hugging Face Models (for demo)
transformers>=4.30.0

# Data Processing
numpy>=1.24.0
pillow>=9.5.0
opencv-python>=4.7.0

# Data Augmentation
albumentations>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Logging and Monitoring
tensorboard>=2.12.0
tqdm>=4.65.0

# Web UI
streamlit>=1.22.0
pandas>=2.0.0

# Configuration and Utilities
pyyaml>=6.0
pathlib>=1.0.1
